Move parsing utilities to utility class
